% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_ncvs.R
\name{get_ncvs}
\alias{get_ncvs}
\title{Download National Crime Victimization Survey (NCVS) data from the Bureau
of Justice Statistics' API.}
\usage{
get_ncvs(filename, path = NULL, survey, year, pop = NULL, ...)
}
\arguments{
\item{filename}{File name, to be downloaded if local file does not exist}

\item{path}{path to local file or alternative download directory}

\item{survey}{survey type string, accepts \code{"personal"} or
\code{"household"}}

\item{year}{survey year in yyyy format}

\item{pop}{logical; \code{TRUE} indicates population data, \code{FALSE}
indicates counts}

\item{...}{additional arguments are passed on to \code{\link[utils]{read.csv}()}
or \code{\link[jsonlite]{fromJSON}()}}
}
\description{
Currently provides support for csv and json file formats.
}

