% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_ncvs.R
\name{get_ncvs}
\alias{get_ncvs}
\title{Download National Crime Victimization Survey (NCVS) data from the Bureau
of Justice Statistics' API.}
\usage{
get_ncvs(survey, year, pop = NULL, filename, path, format = "csv", ...)
}
\arguments{
\item{survey}{survey type; accepts "personal" or "household"}

\item{year}{survey year in YYYY format}

\item{pop}{logical; TRUE indicates population data, FALSE indicates counts}

\item{filename}{name of the downloaded file}

\item{path}{directory to store downloaded file}

\item{format}{file format type; "csv" by default, also accepts "json"}
}
\description{
Download National Crime Victimization Survey (NCVS) data from the Bureau
of Justice Statistics' API.
}

